Format Rust files on save via coc.nvim
rust.vim's autoformatting had the annoying property of making the cursor
move to the first line of the file when doing an undo after saving,
which causes me to lose focus of what I was doing. coc.nvim's
autoformatting doesn't have this.

It might be related to this open issue:
https://github.com/rust-lang/rust.vim/issues/381